Move media types util to core folder
